pyproject.toml
[project]
name = "inthebasement"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discid>=1.3.0",
    "jupyter>=1.1.1",
    "librosa>=0.11.0",
    "matplotlib>=3.10.6",
    "mkdocs>=1.6.1",
    "mkdocs-autorefs>=1.4.3",
    "mkdocs-awesome-nav>=3.2.0",
    "mkdocs-material>=9.6.20",
    "musicbrainzngs>=0.7.1",
    "nbconvert>=7.16.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "psutils>=3.3.14",
    "pyloudnorm>=0.1.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "soundfile>=0.13.1",
    "tabulate>=0.9.0",
]
